// apps/frontend/models/room-booking.model.ts
export interface RoomBooking {
  bookedForId: string;
  feedback: any;
  checkoutTime: string;
  checkinTime: string;
  bookedFor: string;
  cancelReason: string;
  listDevice: any[];
  roomId: string;
  checkinSlotId: string;
  checkoutSlotId: string;
  acceptedBy: string;
  requestedById: string;
  checkinDate: string;
  checkoutSlot: string;
  checkinSlot: string;
  stt: number;
  id: string;
  roomid: string;
  roomName: string
  timeCheckin: string;
  timeCheckout: string;
  requestedBy: string;
  requestedAt: string;
  status: string;
  bookedAt: string;
  checkInAt: string;
  checkOutAt: string;
  updatedAt: string;
  updatedBy: string;
  cancelledBy: string;
  reason: string;
  description: string;
}